+++ src/scout_apm/compat.py
@@ -2,6 +2,7 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 import sys
+import time
 
 string_type = str if sys.version_info[0] >= 3 else basestring  # noqa: F821
 text_type = str if sys.version_info[0] >= 3 else unicode  # noqa: F821
@@ -30,4 +31,23 @@ except ImportError:
     # Python 2.x
     import Queue as queue
 
-__all__ = ["ContextDecorator", "queue", "string_type", "text_type"]
+
+if sys.version_info >= (3, 3):
+
+    def datetime_to_timestamp(datetime_obj):
+        return datetime_obj.timestamp()
+
+
+else:
+
+    def datetime_to_timestamp(datetime_obj):
+        return time.mktime(datetime_obj.timetuple())
+
+
+__all__ = [
+    "ContextDecorator",
+    "datetime_to_timestamp",
+    "queue",
+    "string_type",
+    "text_type",
+]