+++ src/scout_apm/core/util.py
@@ -1,9 +1,38 @@
 # coding=utf-8
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+import time
 
-# Takes an integer or a string containing an integer and returns
-# the octal value.
-# Raises a ValueError if the value cannot be converted to octal.
+# Cutoff epoch is used for determining ambiguous timestamp boundaries, and is
+# just over 10 years ago at time of writing
+CUTOFF_EPOCH_S = time.mktime((2009, 6, 1, 0, 0, 0, 0, 0, 0))
+CUTOFF_EPOCH_MS = CUTOFF_EPOCH_S * 1000.0
+CUTOFF_EPOCH_US = CUTOFF_EPOCH_S * 1000000.0
+CUTOFF_EPOCH_NS = CUTOFF_EPOCH_S * 1000000000.0
+
+
+def convert_ambiguous_timestamp_to_ns(timestamp):
+    """
+    Convert an ambiguous float timestamp that could be in nanoseconds,
+    microseconds, milliseconds, or seconds to nanoseconds. Return 0.0 for
+    values in the more than 10 years ago.
+    """
+    if timestamp > CUTOFF_EPOCH_NS:
+        converted_timestamp = timestamp
+    elif timestamp > CUTOFF_EPOCH_US:
+        converted_timestamp = timestamp * 1000.0
+    elif timestamp > CUTOFF_EPOCH_MS:
+        converted_timestamp = timestamp * 1000000.0
+    elif timestamp > CUTOFF_EPOCH_S:
+        converted_timestamp = timestamp * 1000000000.0
+    else:
+        return 0.0
+    return converted_timestamp
+
+
 def octal(value):
+    """
+    Takes an integer or a string containing an integer and returns the octal
+    value. Raises a ValueError if the value cannot be converted to octal.
+    """
     return int("{}".format(value), 8)