+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Pending
+
+### Added
+
+- Add support to Django integration for tracking request queue timing from the
+  value of the `X-Queue-Start` or `X-Request-Start` header
+
 ## [2.0.5] 2019-06-21
 
 ### Added